feat: 미션 추가 api 구현

+++ backend/choki/src/main/java/com/yeojiphap/choki/domain/mission/domain/Mission.java
@@ -23,7 +23,7 @@ public class Mission {
 
     private String content;
 
-    private int exp;
+    private Long exp;
 
     private Status status;
